Testing & Debugging Professeur Controller

+++ delibrations/src/main/java/ma/enset/delibrations/services/servicesImp/ProfesseurServiceImpl.java
@@ -9,6 +9,7 @@ import ma.enset.delibrations.dtos.responses.ProfesseurResponseDTO;
 import ma.enset.delibrations.entities.Element;
 import ma.enset.delibrations.entities.Professeur;
 import ma.enset.delibrations.exceptions.CannotProceedException;
+import ma.enset.delibrations.exceptions.ElementNotFoundException;
 import ma.enset.delibrations.exceptions.ProfesseurNotFoundException;
 import ma.enset.delibrations.repositories.ElementRepository;
 import ma.enset.delibrations.repositories.ProfesseurRepository;
@@ -28,13 +29,18 @@ public class ProfesseurServiceImpl implements ProfesseurService {
     private ProfesseurRepository professeurRepository;
     private ProfesseurMapper professeurMapper;
     private ElementMapper elementMapper;
-    private ElementRepository elementRepository;
+    private ElementServiceImpl elementService;
 
     @Override
-    public ProfesseurResponseDTO createProfesseur(ProfesseurRequestDTO professeurRequestDTO) throws CannotProceedException {
+    public ProfesseurResponseDTO createProfesseur(ProfesseurRequestDTO professeurRequestDTO) throws CannotProceedException, ProfesseurNotFoundException, ElementNotFoundException {
         if(professeurRequestDTO!=null){
             Professeur professeur = professeurMapper.fromRequestDTOtoEntity(professeurRequestDTO);
             professeur.setCreatedAt(new Date());
+            /*Element*/
+            for (Element element : professeur.getElementModules()) {
+                element.setProfesseur(professeur);
+                elementService.addElement(elementMapper.fromEntitytoRequestDTO(element));
+            }
             professeurRepository.save(professeur);
             return professeurMapper.fromEntityToResponseDTO(professeur);
         }
@@ -42,21 +48,27 @@ public class ProfesseurServiceImpl implements ProfesseurService {
     }
 
     @Override
-    public ProfesseurResponseDTO updateProfesseur(Long id, ProfesseurRequestDTO professeurRequestDTO) throws ProfesseurNotFoundException {
+    public ProfesseurResponseDTO updateProfesseur(Long id, ProfesseurRequestDTO professeurRequestDTO) throws ProfesseurNotFoundException, ElementNotFoundException {
         if (id != null && professeurRequestDTO != null) {
-            Professeur professeur = professeurRepository.findById(id).orElseThrow(() -> new ProfesseurNotFoundException(id));
+            Professeur professeur = professeurRepository.findByIdAndSoftDeleteIsFalse(id);
+            if (professeur == null) throw new ProfesseurNotFoundException(id);
             if (professeurRequestDTO.getCin() != null) professeur.setCin(professeurRequestDTO.getCin());
             if (professeurRequestDTO.getNom() != null) professeur.setNom(professeurRequestDTO.getNom());
             if (professeurRequestDTO.getPrenom() != null) professeur.setPrenom(professeurRequestDTO.getPrenom());
             if (professeurRequestDTO.getEmail() != null) professeur.setEmail(professeurRequestDTO.getEmail());
             if (professeurRequestDTO.getElementModules() != null){
-                professeur.getElementModules().clear();
-                professeurRequestDTO.getElementModules().forEach(elementModuleRequestDTO -> {
-                    Element element = elementMapper.fromRequestDTOtoEntity(elementModuleRequestDTO);
+                Long[] elementModules = professeurRequestDTO.getElementModules();
+                List<Element> elements = new ArrayList<>();
+                for (Long elementId : elementModules) {
+                    Element element = elementService.getElement(elementId);
+                    if (element == null) throw new ElementNotFoundException("Element with id " + elementId + " not found");
+                    elements.add(element);
+                }
+                professeur.setElementModules(elements);
+                for (Element element : elements) {
                     element.setProfesseur(professeur);
-                    elementRepository.save(element);
-                    professeur.getElementModules().add(element);
-                });
+                    elementService.updateElement(elementMapper.fromEntitytoRequestDTO(element));
+                }
             }
             professeur.setUpdatedOn(new Date());
             professeurRepository.save(professeur);
@@ -68,13 +80,14 @@ public class ProfesseurServiceImpl implements ProfesseurService {
     @Override
     public ProfesseurResponseDTO getProfesseur(Long id) throws ProfesseurNotFoundException {
         if (id == null) return null;
-        Professeur professeur = professeurRepository.findById(id).orElseThrow(() -> new ProfesseurNotFoundException(id));
+        Professeur professeur = professeurRepository.findByIdAndSoftDeleteIsFalse(id);
+        if (professeur == null) throw new ProfesseurNotFoundException(id);
         return professeurMapper.fromEntityToResponseDTO(professeur);
     }
 
     @Override
     public List<ProfesseurResponseDTO> getProfesseurs() {
-        List<Professeur> professeurs = professeurRepository.findAll();
+        List<Professeur> professeurs = professeurRepository.findBySoftDeleteIsFalse();
         List<ProfesseurResponseDTO> professeurResponseDTOS = new ArrayList<>();
         professeurs.forEach(professeur -> {
             professeurResponseDTOS.add(professeurMapper.fromEntityToResponseDTO(professeur));
@@ -85,7 +98,12 @@ public class ProfesseurServiceImpl implements ProfesseurService {
     @Override
     public void deleteProfesseur(Long id) throws ProfesseurNotFoundException {
         if (id == null) return;
-        Professeur professeur = professeurRepository.findById(id).orElseThrow(() -> new ProfesseurNotFoundException(id));
+        Professeur professeur = professeurRepository.findByIdAndSoftDeleteIsFalse(id);
+        if (professeur == null) throw new ProfesseurNotFoundException(id);
+        /*Hardcode del
         professeurRepository.delete(professeur);
+         */
+        professeur.setSoftDelete(true);
+        professeurRepository.save(professeur);
     }
 }
